保存版本

+++ examples/21_quant_gemm/quant_gemm.cpp
@@ -73,9 +73,9 @@ void QuantGemm(
     using BType = gemm::MatmulType<int8_t, LayoutB>;
     using CType = gemm::MatmulType<bfloat16_t, LayoutC>;
     using XType = gemm::MatmulType<int32_t, LayoutC>;
-    using ScaleType = gemm::MatmulType<float, LayoutScale>;
-    using PerTokenScaleType = gemm::MatmulType<float, LayoutPerTokenScale>;
-    using BiasType = gemm::MatmulType<float, LayoutBias>;
+    using ScaleType = gemm::MatmulType<bfloat16_t, LayoutScale>;
+    using PerTokenScaleType = gemm::MatmulType<bfloat16_t, LayoutPerTokenScale>;
+    using BiasType = gemm::MatmulType<bfloat16_t, LayoutBias>;
     using TempType = gemm::MatmulType<float, LayoutC>;
     using L1TileShape = MatmulShape<256, 128, 256>;
     using L0TileShape = MatmulShape<256, 128, 128>;
@@ -195,9 +195,9 @@ void Run(Options options){
     size_t sizeB = lenB * sizeof(int8_t);
     size_t sizeC = lenC * sizeof(bfloat16_t);
     size_t sizeX = lenX * sizeof(int32_t);
-    size_t sizeScale = lenScale * sizeof(float);
-    size_t sizePerTokenScale = lenPerTokenScale * sizeof(float);
-    size_t sizeBias = lenBias * sizeof(float);
+    size_t sizeScale = lenScale * sizeof(bfloat16_t);
+    size_t sizePerTokenScale = lenPerTokenScale * sizeof(bfloat16_t);
+    size_t sizeBias = lenBias * sizeof(bfloat16_t);
 
     const uint32_t align = 256;
     using LayoutA = layout::RowMajor;
@@ -219,9 +219,9 @@ void Run(Options options){
 
     std::vector<int8_t> hostA(lenA);
     std::vector<int8_t> hostB(lenB);
-    std::vector<float> hostScale(lenScale);
-    std::vector<float> hostPerTokenScale(lenPerTokenScale);
-    std::vector<float> hostBias(lenBias);
+    std::vector<bfloat16> hostScale(lenScale);
+    std::vector<bfloat16> hostPerTokenScale(lenPerTokenScale);
+    std::vector<bfloat16> hostBias(lenBias);
     golden::FillRandomData(hostA,  -16, 16);
     golden::FillRandomData(hostB,  -16, 16);
     golden::FillRandomData(hostScale, 0.0, 1.0);
@@ -246,15 +246,15 @@ void Run(Options options){
     } else {
         ACL_CHECK(aclrtMalloc(reinterpret_cast<void **>(&deviceWB), sizeWB, ACL_MEM_MALLOC_HUGE_FIRST));
     }
-    float* deviceScale{nullptr};
+    bfloat16* deviceScale{nullptr};
     ACL_CHECK(aclrtMalloc(reinterpret_cast<void **>(&deviceScale), sizeScale, ACL_MEM_MALLOC_HUGE_FIRST));
     ACL_CHECK(aclrtMemcpy(deviceScale, sizeScale, hostScale.data(), sizeScale, ACL_MEMCPY_HOST_TO_DEVICE));
 
-    float* devicePerTokenScale{nullptr};
+    bfloat16* devicePerTokenScale{nullptr};
     ACL_CHECK(aclrtMalloc(reinterpret_cast<void **>(&devicePerTokenScale), sizePerTokenScale, ACL_MEM_MALLOC_HUGE_FIRST));
     ACL_CHECK(aclrtMemcpy(devicePerTokenScale, sizePerTokenScale, hostPerTokenScale.data(), sizePerTokenScale, ACL_MEMCPY_HOST_TO_DEVICE));
 
-    float* deviceBias{nullptr};
+    bfloat16* deviceBias{nullptr};
     ACL_CHECK(aclrtMalloc(reinterpret_cast<void **>(&deviceBias), sizeBias, ACL_MEM_MALLOC_HUGE_FIRST));
     ACL_CHECK(aclrtMemcpy(deviceBias, sizeBias, hostBias.data(), sizeBias, ACL_MEMCPY_HOST_TO_DEVICE));
 
@@ -298,6 +298,7 @@ void Run(Options options){
         hostPerTokenScale, layoutPerTokenScale,
         hostBias, layoutBias,
         hostGolden, layoutC);
+    // golden::ComputeMatmul(options.problemShape, hostA, layoutA, hostB, layoutB, hostGolden, layoutC);
     std::vector<uint64_t> errorIndices = golden::CompareData(hostRes, hostGolden, m * n);
     if (errorIndices.empty()) {
         std::cout << "Compare success." << std::endl;
